requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow==2.19.0
gdown
google-generativeai